util/converter/pyproject.toml
[tool.poetry]
name = "reynard_converter"
version = "0.1.0"
description = ""
authors = ["Delano Flipse <delanoflipse@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
argparse = "^1.4.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
